% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{evs_values}
\alias{evs_values}
\title{EVS Values}
\format{
A data frame with columns:
\describe{
  \item{variable_name}{The variable name from EVS}
  \item{evs_question_label}{Question text from EVS}
  \item{evs_value_label}{Label for the value in EVS}
}
}
\usage{
data(evs_values)
}
\description{
This dataset contains extracted values from the European Values Study (EVS).
}
\examples{
data(evs_values)
head(evs_values)
}
\keyword{datasets}
