% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivs_creator.R
\name{ivs_creator}
\alias{ivs_creator}
\title{Create Integrated Values Study Dataset}
\usage{
ivs_creator(wvs_path, evs_path)
}
\arguments{
\item{wvs_path}{Path to the WVS .dta file.}

\item{evs_path}{Path to the EVS .dta file.}
}
\value{
A merged tibble containing harmonized data from WVS and EVS.

A merged tibble containing harmonized data from WVS and EVS.
}
\description{
This function reads and processes World Values Survey (WVS) and European Values Study (EVS) data,
harmonizing common variables and merging unique ones.
}
