% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ivs_values}
\alias{ivs_values}
\title{IVS Values Comparison}
\format{
A data frame with columns:
\describe{
  \item{variable_name}{The variable name}
  \item{wvs_question_label}{Question text from WVS}
  \item{evs_question_label}{Question text from EVS}
  \item{wvs_value_label}{Label for the value in WVS}
  \item{evs_value_label}{Label for the value in EVS}
  \item{wvs_only}{1 if only in WVS, else 0}
  \item{evs_only}{1 if only in EVS, else 0}
  \item{wvs_evs_value_conflict}{1 if WVS & EVS values conflict, else 0}
}
}
\usage{
data(ivs_values)
}
\description{
Merged dataset comparing WVS and EVS values.
}
\examples{
data(ivs_values)
head(ivs_values)
}
\keyword{datasets}
