% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_whogov.R
\name{merge_whogov}
\alias{merge_whogov}
\title{Merge WVS Data with WhoGov Party IDs}
\usage{
merge_whogov(wvs)
}
\arguments{
\item{wvs}{A dataframe containing WVS data with columns `first_vote_partyfacts_id`,
`second_vote_partyfacts_id`, and `never_vote_partyfacts_id`.}
}
\value{
A dataframe with merged WhoGov IDs.
}
\description{
This function merges WVS vote choices with Partyfacts and WhoGov IDs.
It assumes that the `merge_wvs_party()` function has been run beforehand.
}
