% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_wvs_party_id.R
\name{merge_wvs_party_id}
\alias{merge_wvs_party_id}
\title{Merge WVS Party ID Data}
\usage{
merge_wvs_party_id(wvs)
}
\arguments{
\item{wvs}{A dataframe containing WVS data with columns `E179WVS`, `E180WVS`, `E182`.}
}
\value{
A dataframe with merged party facts IDs.
}
\description{
This function merges World Values Survey (WVS) vote choices with party IDs. It appends three variables to the dataset. first_vote_partyfacts_id, second_vote_partyfacts_id, never_vote_partyfacts_id
}
\examples{
wvs_test <- data.frame(E179WVS = c(101, 102), E180WVS = c(103, 104), E182 = c(105, 106))
merged_data <- merge_wvs_party_id(wvs_test)
head(merged_data)
}
