% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codebook.R
\name{wvs_items}
\alias{wvs_items}
\title{Retrieve human-readable variable label(s)}
\usage{
wvs_items(
  vars = NULL,
  group = NULL,
  columns = c("label", "group", "direction", "min", "max")
)
}
\arguments{
\item{vars}{Optional character vector of variable names to include.}

\item{group}{Optional character vector of groups to include.}

\item{columns}{Optional character vector of columns to return. Valid
values are \code{"label"}, \code{"group"}, \code{"direction"}, \code{"min"}, and \code{"max"}.}

\item{v}{Character vector of variable names.
List item metadata from the static EVS/WVS codebook}
}
\value{
Data.frame describing item metadata. Row names are variable names.
}
\description{
Retrieve human-readable variable label(s)
}
