% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wvs_values}
\alias{wvs_values}
\title{WVS Values}
\format{
A data frame with columns:
\describe{
  \item{variable_name}{The variable name from WVS}
  \item{wvs_question_label}{Question text from WVS}
  \item{wvs_value_label}{Label for the value in WVS}
}
}
\usage{
data(wvs_values)
}
\description{
Extracted values from the World Values Survey (WVS).
}
\examples{
data(wvs_values)
head(wvs_values)
}
\keyword{datasets}
